refactor: Remove unused Queue model and update music stop command
This commit removes the unused Queue model import from the music stop command file. It also updates the logic in the music stop command to delete the queue using the queue.delete() method instead of the deprecated Queue.deleteOne() method. This change improves the efficiency and maintainability of the code.